requirements.txt
streamlit
google-generativeai
gtts
streamlit-mic-recorder